DiscordBot/requirements.txt
discord.py>=2.0.0
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
python-dotenv>=0.19.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
seaborn>=0.11.0
matplotlib>=3.4.0 